세션 방식 로그인 적용

+++ src/main/java/com/team/chatproject/domain/Member.java
@@ -3,11 +3,9 @@ package com.team.chatproject.domain;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.security.core.GrantedAuthority;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @Getter
@@ -16,7 +14,7 @@ import java.util.List;
 public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @Column(unique = true)
     private String loginId ;
@@ -38,5 +36,6 @@ public class Member {
 
     private int delStatus;
     private LocalDateTime delDate;
+
 
 }
